+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 angular-leaflet-directive changelog
 ===================================
 
+## angular-leaflet-directive (trunk)
+* Splitted the project in multiple sub-directives, children of the main "<leaflet>" directive.
+* Added a leafletData service, where all the registered Leaflet object will be stored.
+* Added a leafletHelpers factory.
+
 ## angular-leaflet-directive (0.5.1)
 * fix: wrap marker click event broadcast in scope apply